Move decompose_word into from constants::util into gadget::utilities.
This helper is not used outside of the gadget.
